// packages/docx/src/arabic-joining.generated.ts
export const JT_NAMES = [
  'U', 'C', 'D', 'L', 'R', 'T',
] as const;
// ...
export const JT_RANGE_VALUE: number[] = [
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 2, 0, 4, 2, 4, 2, 4, 2, 4, 2, 1,
  2, 4, 2, 5, 0, 2, 5, 4, 0, 4, 2, 4, 2, 4, 2, 4,
  2, 4, 2, 4, 2, 4, 0, 4, 5, 0, 5, 0, 5, 0, 5, 4,
  0, 2, 0, 2, 0, 5, 4, 5, 2, 4, 2, 4, 2, 4, 2, 4,
  2, 4, 2, 4, 5, 0, 4, 2, 4, 2, 4, 2, 4, 2, 4, 2,
  4, 2, 0, 5, 0, 2, 5, 0, 1, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 4, 2, 4, 2, 4, 2, 4, 2, 4, 5, 0, 2,
  0, 2, 0, 4, 2, 4, 0, 4, 1, 2, 0, 2, 4, 2, 0, 5,
  2, 4, 0, 4, 2, 4, 2, 4, 2, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  2, 0, 1, 5, 0, 5, 0, 2, 0, 5, 2, 5, 2, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 1, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 2, 3, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0,
  5, 0, 5, 0, 2, 4, 0, 4, 0, 4, 0, 3, 4, 2, 3, 2,
  4, 2, 4, 0, 4, 5, 0, 2, 4, 0, 2, 4, 2, 4, 2, 4,
  2, 4, 2, 4, 2, 4, 0, 4, 2, 0, 3, 2, 4, 2, 5, 0,
  5, 0, 5, 0, 4, 2, 0, 2, 0, 5, 0, 2, 4, 2, 0, 5,
  2, 4, 0, 2, 4, 2, 5, 0, 2, 0, 2, 4, 0, 2, 4, 2,
  4, 2, 0, 2, 4, 2, 0, 4, 2, 3, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 5,
  0, 5, 0, 5, 0, 5, 0, 5, 0, 5, 0, 2, 5, 0, 5, 0,
  5, 0, 5, 0,
];

// Generated from UCD ArabicShaping-17.0.0.txt (dated 2025-08-14).
// Joining_Group families follow harfbuzz-arabic.c getArabicProperties' group-to-class mapping.
export const KASHIDA_SEEN_PREV: number[] = [
  0x633, 0x634, 0x635, 0x636, 0x69A, 0x69B, 0x69C, 0x69D, 0x69E, 0x6FA, 0x6FB, 0x75C,
  0x76D, 0x770, 0x77D, 0x77E, 0x8AF,
];

export const KASHIDA_HAH_PREV: number[] = [
  0x62C, 0x62D, 0x62E, 0x681, 0x682, 0x683, 0x684, 0x685, 0x686, 0x687, 0x6BF, 0x757,
  0x758, 0x76E, 0x76F, 0x772, 0x77C, 0x88A, 0x8A2, 0x8C1, 0x8C5, 0x8C6,
];

export const KASHIDA_BEH_PREV: number[] = [
  0x628, 0x62A, 0x62B, 0x66E, 0x679, 0x67A, 0x67B, 0x67C, 0x67D, 0x67E, 0x67F, 0x680,
  0x750, 0x751, 0x752, 0x753, 0x754, 0x755, 0x756, 0x8A0, 0x8A1, 0x8B6, 0x8B7, 0x8B8,
  0x8BE, 0x8BF, 0x8C0,
];

export const KASHIDA_REH_YEH_CUR: number[] = [
  0x626, 0x631, 0x632, 0x63D, 0x63E, 0x63F, 0x649, 0x64A, 0x678, 0x691, 0x692, 0x693,
  0x694, 0x695, 0x696, 0x697, 0x698, 0x699, 0x6CC, 0x6CE, 0x6D0, 0x6D1, 0x6EF, 0x75B,
  0x76B, 0x76C, 0x771, 0x775, 0x776, 0x777, 0x8A8, 0x8A9, 0x8AA, 0x8B2, 0x8B9, 0x8BA,
  0x10EC7,
];

export const KASHIDA_ALEF_TAH_CUR: number[] = [
  0x622, 0x623, 0x625, 0x627, 0x637, 0x638, 0x63B, 0x63C, 0x643, 0x644, 0x671, 0x672,
  0x673, 0x675, 0x69F, 0x6A9, 0x6AB, 0x6AC, 0x6AD, 0x6AE, 0x6AF, 0x6B0, 0x6B1, 0x6B2,
  0x6B3, 0x6B4, 0x6B5, 0x6B6, 0x6B7, 0x6B8, 0x762, 0x763, 0x764, 0x76A, 0x773, 0x774,
  0x77F, 0x870, 0x871, 0x872, 0x873, 0x874, 0x875, 0x876, 0x877, 0x878, 0x879, 0x87A,
  0x87B, 0x87C, 0x87D, 0x87E, 0x87F, 0x880, 0x881, 0x882, 0x88B, 0x88C, 0x88D, 0x8A3,
  0x8A6, 0x8B0, 0x8B4, 0x8C2, 0x8C7, 0x8C8, 0x10EC3, 0x10EC4,
];

export const KASHIDA_WAW_AIN_CUR: number[] = [
  0x624, 0x639, 0x63A, 0x641, 0x642, 0x648, 0x66F, 0x676, 0x677, 0x6A0, 0x6A1, 0x6A2,
  0x6A3, 0x6A4, 0x6A5, 0x6A6, 0x6A7, 0x6A8, 0x6C4, 0x6C5, 0x6C6, 0x6C7, 0x6C8, 0x6C9,
  0x6CA, 0x6CB, 0x6CF, 0x6FC, 0x75D, 0x75E, 0x75F, 0x760, 0x761, 0x778, 0x779, 0x8A4,
  0x8A5, 0x8AB, 0x8B3, 0x8B5, 0x8C3,
];